Moved errors to own folder.
