+++ src/errors/PathfindingInterupptedError.ts
