+++ src/errors/UnknownCollectableError.ts
